requirements.txt
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
arxiv==2.0.0
asyncio==3.4.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cssselect2==0.7.0
dataclasses-json==0.6.2
Deprecated==1.2.14
distro==1.8.0
docopt==0.6.2
duckduckgo-search==3.9.5
emoji==2.9.0
faiss-cpu==1.7.4
fastapi==0.104.1
feedparser==6.0.10
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
gpt4all==2.0.2
greenlet==3.0.0
grpcio==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kombu==5.3.4
kubernetes==28.1.0
langchain==0.0.335
langchain-experimental==0.0.42
langchainhub==0.1.14
langdetect==1.0.9
langsmith==0.0.66
lxml==4.9.3
Markdown==3.5.1
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.20.1
md2pdf==1.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.2.4
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
permchain==0.0.3
Pillow==10.1.0
playwright==1.39.0
posthog==3.0.2
prompt-toolkit==3.0.43
protobuf==4.25.1
psycopg2-binary==2.9.9
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydyf==0.8.0
pyee==11.0.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pypdf==3.17.2
pyphen==0.14.0
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.1
rapidfuzz==3.5.2
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
selenium==4.15.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tavily-python==0.2.8
tenacity==8.2.3
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.35.2
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
types-click==7.1.8
types-Flask==1.1.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-requests==2.31.0.6
types-urllib3==1.26.25.14
types-Werkzeug==1.0.9
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.11.4
unstructured-client==0.15.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
weasyprint==60.1
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.3
zipp==3.17.0
zopfli==0.2.3
